refactor: useStakingInput helpers to own file
