+++ src/components/common/ColonyActions/ActionDetailsPage/DefaultMotion/MotionPhaseWidget/StakingWidget/index.ts
@@ -3,8 +3,8 @@ export { default as SingleTotalStake } from './SingleTotalStake';
 export {
   default as StakingInput,
   StakingSlider,
-  getStakingTransformFn,
   StakingWidgetValues,
+  SLIDER_AMOUNT_KEY,
 } from './StakingInput';
 export { default as GroupedTotalStake } from './GroupedTotalStake';
 export * from './StakingWidgetProvider';
