+++ src/hooks/helpers/motionHookHelpers.ts
@@ -1,0 +1,30 @@
+import { BigNumber } from 'ethers';
+import { getStakeFromSlider } from '~common/ColonyActions/ActionDetailsPage/DefaultMotion/MotionPhaseWidget/StakingWidget';
+import { Action, ActionTypes } from '~redux';
+import { mapPayload } from '~utils/actions';
+
+type StakeMotionPayload = Action<ActionTypes.MOTION_STAKE>['payload'];
+
+export const getStakingTransformFn = (
+  remainingToStake: string,
+  userMinStake: string,
+  userAddress: string,
+  colonyAddress: string,
+  motionId: string,
+  vote: number,
+) =>
+  mapPayload(({ amount: sliderAmount }) => {
+    const finalStake = getStakeFromSlider(
+      sliderAmount,
+      remainingToStake,
+      userMinStake,
+    );
+
+    return {
+      amount: finalStake,
+      userAddress,
+      colonyAddress,
+      motionId: BigNumber.from(motionId),
+      vote,
+    } as StakeMotionPayload;
+  });
