+++ src/hooks/useObjectButton.ts
@@ -1,14 +1,16 @@
 import { useFormContext } from 'react-hook-form';
 
 import {
-  getStakingTransformFn,
   useStakingWidgetContext,
+  SLIDER_AMOUNT_KEY,
+  StakingWidgetValues,
 } from '~common/ColonyActions/ActionDetailsPage/DefaultMotion/MotionPhaseWidget/StakingWidget';
-import { SLIDER_AMOUNT_KEY } from '~common/ColonyActions/ActionDetailsPage/DefaultMotion/MotionPhaseWidget/StakingWidget/StakingInput';
 import { RaiseObjectionDialog } from '~common/Dialogs';
 import { useDialog } from '~shared/Dialog';
+import { OnSuccess } from '~shared/Fields';
 import { MotionVote } from '~utils/colonyMotions';
 
+import { getStakingTransformFn } from './helpers';
 import useAppContext from './useAppContext';
 import useColonyContext from './useColonyContext';
 
@@ -35,10 +37,14 @@ const useObjectButton = () => {
 
   const canUserStakedNay = !!(user && nayRemaining !== '0');
 
+  const handleStakeSuccess: OnSuccess<StakingWidgetValues> = (_, { reset }) => {
+    reset();
+  };
   const handleObjection = () =>
     openRaiseObjectionDialog({
       canBeStaked: canUserStakedNay,
       transform,
+      handleStakeSuccess,
       setIsSummary,
       amount: getValues(SLIDER_AMOUNT_KEY),
     });
